Create human profile database from bloodmeal profiles

Create human profile database from bloodmeal profiles

Usage

create_db_from_bloodmeals(
  bloodmeal_profiles,
  kit,
  peak_thresh,
  rm_markers = c("AMEL")
)

Arguments

bloodmeal_profiles

Tibble or data frame with alleles for all bloodmeals in reference database including 4 columns: SampleName, Marker, Allele, Height. Height must be numeric or coercible to numeric.

kit

STR kit name from euroformix. To see a list of all kits embedded in euroformix use euroformix::getKit(). If your kit is not included, see vignette("bistro") for details on how to include your own kit.

peak_thresh

Allele peak height threshold in RFUs. All peaks under this threshold will be filtered out. If prior filtering was performed, this number should be equal to or greater than that number. Also used for threshT argument in euroformix::contLikSearch().

rm_markers

A vector indicating what markers should be removed prior to calculating log10LRs. NULL to include all markers. By default, for the bistro function AMEL is removed as it is not standard to include it in LR calculations.

Value

Human database created from complete single-source bloodmeals. Complete is defined as the number of markers in the kit minus the number of markers in rm_markers.

Examples

if (FALSE) { # \dontrun{
# load example data
path_to_data <- paste0(
  "https://raw.githubusercontent.com/duke-malaria-collaboratory/",
  "bistro_validation/main/data/provedit/provedit_samples_mass200thresh.csv"
)
samples <- readr::read_csv(path_to_data)
create_db_from_bloodmeals(samples, kit = "identifiler", peak_thresh = 200)
} # }