Identify matches for multiple bloodmeal-human pairs

Identify matches for multiple bloodmeal-human pairs

Usage

identify_matches(log10_lrs, bloodmeal_ids = NULL, check_inputs = TRUE)

Arguments

log10_lrs

Output from calc_log10_lrs() or from bistro with return_lrs = TRUE (lrs: the second element in the list)

bloodmeal_ids

Vector of bloodmeal ids from the SampleName column in bloodmeal_profiles for which to compute log10_lrs. If NULL, all ids in the input dataframe will be used. Default: NULL

check_inputs

A boolean indicating whether or not to check the inputs to the function. Default: TRUE

Value

A tibble with the same output as for bistro().

Examples

bm_profs <- prep_bloodmeal_profiles(bloodmeal_profiles, peak_thresh = 200)
#> Removing 6 peaks under the threshold of 200 RFU.
#> For 1/4 bloodmeal ids, all peaks are below the threshold
hu_profs <- prep_human_profiles(human_profiles)
log10_lrs <- calc_log10_lrs(bm_profs,
  hu_profs,
  bloodmeal_ids = "evid1",
  pop_allele_freqs = pop_allele_freqs,
  kit = "ESX17",
  peak_thresh = 200
)
#> 1/17 markers in kit but not in bloodmeal_profiles: AMEL
#> 1/17 markers in kit but not in human_profiles: AMEL
#> # bloodmeal ids: 1
#> # human ids: 3
#> Bloodmeal id 1/1
#> Human id 1/3
#> Human id 2/3
#> Human id 3/3
matches <- identify_matches(log10_lrs)