Check bistro inputs

Check bistro inputs

Usage

check_bistro_inputs(
  bloodmeal_profiles,
  human_profiles,
  kit,
  peak_thresh,
  pop_allele_freqs,
  calc_allele_freqs,
  bloodmeal_ids,
  human_ids,
  rm_twins,
  rm_markers,
  model_degrad,
  model_bw_stutt,
  model_fw_stutt,
  difftol,
  threads,
  seed,
  time_limit,
  return_lrs
)

Arguments

bloodmeal_profiles

Tibble or data frame with alleles for all bloodmeals in reference database including 4 columns: SampleName, Marker, Allele, Height. Height must be numeric or coercible to numeric.

human_profiles

Tibble or data frame with alleles for all humans in reference database including three columns: SampleName, Marker, Allele.

kit

STR kit name from euroformix. To see a list of all kits embedded in euroformix use euroformix::getKit(). If your kit is not included, see vignette("bistro") for details on how to include your own kit.

peak_thresh

Allele peak height threshold in RFUs. All peaks under this threshold will be filtered out. If prior filtering was performed, this number should be equal to or greater than that number. Also used for threshT argument in euroformix::contLikSearch().

pop_allele_freqs

Tibble or data frame where the first column is the STR allele and the following columns are the frequency of that allele for different markers. Alleles that do not exist for a given marker are coded as NA. If NULL and calc_allele_freqs = TRUE, then population allele frequencies will be calculated from human_profiles.

calc_allele_freqs

A boolean indicating whether or not to calculate allele frequencies from human_profiles. If FALSE, a pop_allele_freqs input is required. Default: FALSE

bloodmeal_ids

Vector of bloodmeal ids from the SampleName column in bloodmeal_profiles for which to compute log10_lrs. If NULL, all ids in the input dataframe will be used. Default: NULL

human_ids

Vector of human ids from the SampleName column in human_profiles for which to compute log10_lrs. If NULL, all ids in the input dataframe will be used. Default: NULL

rm_twins

A boolean indicating whether or not to remove likely twins (identical STR profiles) from the human database prior to identifying matches. Default: TRUE

rm_markers

A vector indicating what markers should be removed prior to calculating log10LRs. NULL to include all markers. By default, for the bistro function AMEL is removed as it is not standard to include it in LR calculations.

model_degrad

A boolean indicating whether or not to model peak degradation. Used for modelDegrad argument in euroformix::contLikSearch(). Default: TRUE

model_bw_stutt

A boolean indicating whether or not to model peak backward stutter. Used for modelBWstutt argument in euroformix::contLikSearch(). Default: FALSE

model_fw_stutt

A boolean indicating whether or not to model peak forward stutter. Used for modelFWstutt argument in euroformix::contLikSearch(). Default: FALSE

difftol

Tolerance for difference in log likelihoods across 2 iterations. euroformix::contLikSearch() argument. Default: 1

threads

Number of threads to use when calculating log10_lrs. euroformix::contLikSearch() argument. Default: 4

seed

Seed when calculating log10_lrs. euroformix::contLikSearch() argument. Default: NULL (no seed)

time_limit

Time limit in minutes to run the euroformix::contLikSearch() function on 1 bloodmeal-human pair. Default: 3

return_lrs

A boolean indicating whether or not to return log10LRs for all bloodmeal-human pairs. Default: FALSE

Value

Error or nothing